+++ ntex/tests/web_ws.rs
@@ -1,0 +1,66 @@
+use std::io;
+
+use bytes::Bytes;
+use futures::{SinkExt, StreamExt};
+
+use ntex::service::{fn_factory_with_config, fn_service};
+use ntex::web::{self, test, ws, App, HttpRequest};
+
+async fn service(msg: ws::Frame) -> Result<Option<ws::Message>, io::Error> {
+    let msg = match msg {
+        ws::Frame::Ping(msg) => ws::Message::Pong(msg),
+        ws::Frame::Text(text) => {
+            ws::Message::Text(String::from_utf8_lossy(&text).to_string())
+        }
+        ws::Frame::Binary(bin) => ws::Message::Binary(bin),
+        ws::Frame::Close(reason) => ws::Message::Close(reason),
+        _ => panic!(),
+    };
+    Ok(Some(msg))
+}
+
+#[ntex::test]
+async fn web_ws() {
+    let srv = test::server(|| {
+        App::new().service(web::resource("/").route(web::to(
+            |req: HttpRequest, pl: web::types::Payload| async move {
+                ws::start::<_, _, _, web::Error>(
+                    fn_factory_with_config(|_| async {
+                        Ok::<_, web::Error>(fn_service(service))
+                    }),
+                    req,
+                    pl,
+                )
+                .await
+            },
+        )))
+    });
+
+    // client service
+    let mut framed = srv.ws().await.unwrap();
+    framed
+        .send(ws::Message::Text("text".to_string()))
+        .await
+        .unwrap();
+    let item = framed.next().await.unwrap().unwrap();
+    assert_eq!(item, ws::Frame::Text(Bytes::from_static(b"text")));
+
+    framed
+        .send(ws::Message::Binary("text".into()))
+        .await
+        .unwrap();
+    let item = framed.next().await.unwrap().unwrap();
+    assert_eq!(item, ws::Frame::Binary(Bytes::from_static(b"text")));
+
+    framed.send(ws::Message::Ping("text".into())).await.unwrap();
+    let item = framed.next().await.unwrap().unwrap();
+    assert_eq!(item, ws::Frame::Pong("text".to_string().into()));
+
+    framed
+        .send(ws::Message::Close(Some(ws::CloseCode::Normal.into())))
+        .await
+        .unwrap();
+
+    let item = framed.next().await.unwrap().unwrap();
+    assert_eq!(item, ws::Frame::Close(Some(ws::CloseCode::Normal.into())));
+}