+++ ntex/CHANGES.md
@@ -10,6 +10,8 @@
 
 * ntex::ws: Add stream and sink wrappers for ws protocol
 
+* ntex::web: Add websockets helper
+
 ## [0.1.14] - 2020-04-27
 
 * ntex::http: Stop client connections pool support future